+++ src/js/components/address/PageLocker.vue
@@ -37,7 +37,7 @@
                         <div class="card-row__name" v-text="$t('address.info.balance')" />
                         <div class="card-row__value" v-if="wallet.balance == '0' || wallet.balance">
                             {{ $ton(wallet.balance) }}
-                            <span v-text="addressMeta.tonIcon || 'TON'" title="TON" />
+                            <span v-text="addressMeta.tonIcon || 'ION'" title="ION" />
 
                             <template v-if="wallet.balance != '0'">
                                 <span v-if="$store.state.exchangeRate" style="color: #717579">
@@ -46,7 +46,7 @@
                             </template>
                         </div>
                         <div v-else class="card-row__value">
-                            <span class="skeleton">00000 TON ≈ 00000 USD</span>
+                            <span class="skeleton">00000 ION ≈ 00000 USD</span>
                         </div>
                     </div>
 
@@ -65,9 +65,9 @@
                             <div class="card-row">
                                 <div class="card-row__name" v-text="$t('locker.total_deposits')" />
                                 <div class="card-row__value">
-                                    <span v-if="locker.total_coins_locked === undefined" class="skeleton">10000 TON</span>
+                                    <span v-if="locker.total_coins_locked === undefined" class="skeleton">10000 ION</span>
                                     <template v-else>
-                                        {{ $ton(locker.total_coins_locked) }} TON
+                                        {{ $ton(locker.total_coins_locked) }} ION
                                     </template>
                                 </div>
                             </div>
@@ -75,9 +75,9 @@
                             <div class="card-row">
                                 <div class="card-row__name" v-text="$t('locker.total_rewards')" />
                                 <div class="card-row__value">
-                                    <span v-if="locker.total_reward === undefined" class="skeleton">10000 TON</span>
+                                    <span v-if="locker.total_reward === undefined" class="skeleton">10000 ION</span>
                                     <template v-else>
-                                        {{ $ton(locker.total_reward) }} TON
+                                        {{ $ton(locker.total_reward) }} ION
                                     </template>
                                 </div>
                             </div>