+++ src/js/components/stats/ChartStaking.vue
@@ -183,7 +183,7 @@ export default {
                 type: 'bar',
                 yAxisID: 'volume',
                 label: this.$t('stats.staking'),
-                suffix: ' TON',
+                suffix: ' ION',
                 parsing: false,
             });